requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
black
scikit-learn
joblib
matplotlib
wordcloud
xhtml2pdf

